+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ## [1.11.11] - unreleased
 ### Added
 - GHORG_GITHUB_REPO_LIST_CONCURRENCY to prevent github secondary ratelimiting
+- Automatic git clone retries (up to 3 retries with 1s, 2s, then 3s delay) on failure
 ### Changed
 ### Deprecated
 ### Removed